Moved structure into appropriate files
